requirements.txt
uvicorn~=0.34.2
fastapi~=0.115.12
pydantic~=2.11.3
SQLAlchemy~=2.0.41
starlette~=0.46.2
aiofiles
pyjwt
jinja2
